requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.1
config==0.5.0.post0
confusable-homoglyphs==3.2.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-bootstrap==0.2.4
django-countries==6.1.2
django-crispy-forms==1.9.1
django-generic-ratings==0.7
django-heroku==0.3.1
django-registration==3.1
django-registration-redux==2.0
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
importlib-resources==1.5.0
Pillow==7.1.2
pipenv==2020.6.2
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.21
virtualenv-clone==0.5.4
whitenoise==5.1.0
zipp==3.1.0